refactor helper functions for EMEP reader into new module and add more tests
